Add build-kubecfg
This commit adds a command to build a kubecfg that optionally can use
aws-iam-authenticator, given we have the
`KOPS_AWS_IAM_AUTHENTICATOR_ENABLED` set to true

We want to enable RBAC using aws-iam-authenticator to auth users against
the Kops cluster via IAM roles.

Note that we have to fetch the CA data using `openssl s_client` as we
don’t have a kubecfg to pull it from yet and we don't want to have to
access the Kops state store S3 bucket where this information is stored.

The build-kubecfg command has a check_env_vars function so that we don't
have to have already declared the cluster_ca function, which keeps
things a little cleaner.

`kopsctl cluster kubecfg build` should be run after assuming an AWS role
within Geodesic.