fix: handling of user+group creation on Debian
The adduser/addgroup commands on Debian based systems are much more
pedantic than other platforms and will error out on some surprising
conditions. Further more, we want to support situations in which the
username or groupname is pulled from systems which support characters
such a spaces in the names which are prohibited in the POSIX
passwd/group files.